Handling for non-unicode characters
When using the macro on platforms whose system uses character sets that are not unicode, string handling fumbles in various locations, most notably when building the export filesnames. This version fixes the file naming such that exporting can work properly. It is unknown whether there are other areas subject to this issue, but it is possible.